[#10] refactor: 인증된 사용자와 익명 사용자 댓글 서비스 다시 통합
