+++ src/main/java/com/mallang/comment/application/CommentService.java
@@ -1,5 +1,6 @@
 package com.mallang.comment.application;
 
+import com.mallang.comment.application.command.WriteAnonymousCommentCommand;
 import com.mallang.comment.application.command.WriteAuthenticatedCommentCommand;
 import com.mallang.comment.domain.Comment;
 import com.mallang.comment.domain.CommentRepository;
@@ -24,6 +25,13 @@ public class CommentService {
         Post post = postRepository.getById(command.postId());
         AuthenticatedWriter writer = authenticatedWriterRepository.getByMemberId(command.memberId());
         Comment comment = command.toComment(post, writer);
+        Comment saved = commentRepository.save(comment);
+        return saved.getId();
+    }
+
+    public Long anonymousWrite(WriteAnonymousCommentCommand command) {
+        Post post = postRepository.getById(command.postId());
+        Comment comment = command.toComment(post);
         Comment saved = commentRepository.save(comment);
         return saved.getId();
     }
